refactoring, removed clamp

+++ benchmark0/prepost/exception.rs
@@ -10,7 +10,7 @@ pub fn first_negative(v: &[i32]) -> (result: i32)
         v.len() <= i32::MAX as usize,
     ensures
         result == -1 || (0 <= result as int && (result as int) < v@.len() && v@[result as int] < 0),
-        result >= 0 ==> forall|j: int| 0 <= j && j < result ==> v@[j] >= 0,
+        result >= 0 ==> forall|j: int| 0 <= j && j < result as int ==> v@[j] >= 0,
         result == -1 ==> forall|j: int| 0 <= j && j < v@.len() ==> v@[j] >= 0,
 {
     assume(false);
